[agent_builder/settings] - refactor: overhaul AgentAccessPublicationDialog with Sheets and Editors management
 - Replace Dialog components with Sheet components for the Editors settings interface
 - Implement use of DataTable to manage the editor's list, with add and remove functionality
 - Add local state management for editors with unsaved changes detection
 - Remove separate EditorsTab component in favor of integrated editor functionality within the main dialog component

[agent_builder/settings] - fix: adjust size for "Add" button in TagsSelector

 - Standardize the "Add" button size to "sm" for consistency within the TagsSelector component
